Several Additions. Good Progress
Receive list of players and community cards. DeterminePlayerHands takes a player and the community cards, combines them, and sorts them by rank. FlusherFinder determines it the list of combined cards contains a flush, royal flush, or straight flush. Need to implement the straight flush part. Need to implement what happens if the flush is neither a royal or a straight.