Added MultipleFinder Func. Updated Testing.
Replaced winningHand field in TestObject with a listt of winningHands. Added a ConvertJson function to Testing to convert the objects in the json tests to another. Updated PerformFinderTest to account for multiple possible winning hands in testing; tests pass like normal with new changes. Added MultipleFinder method to the Algo to recognize Four of a Kind hands. Added log of expected vs actual WinnerHand object information to failed tests. Finished MultipleFinder Method to recognize all remaining hand types. PerformFinderTest now checks if suit verification is necessary or not based on the expected winning hand type.