Restructured the code. Added HandEvaluator class.
HandEvaluator class is now responsible for determining the winning hand given a list of cards for either a player or the community. I combined the Algo class back into one file since it is not as large. Player hand cards have been renamed to HoleCards. Card Value has been renamed to Rank.