Refactor. SNAFU
HandEvaluator class is now responsible for determining the winning hand given a list of cards for either a player or the community. I combined the Algo class back into one file since it is not as large. Player hand cards have been renamed to HoleCards. Card Value has been renamed to Rank. Added ChanceCalculator that uses a Monte Carlo Simulation to calculate the chances of a player winning. I'll add this here because I can, I realize now that there is no need to find the winning hand in the community; When I designed the original structure of the algorithm I was wrong about how the winner in texas hold em is determined; So I'll be reworking this tomorrow because I'm tired as hell; I don't even know how many hours I've put into this by now. It's been like 2 months.